Make the linter pass so the rest of CI can run
Both CI jobs were failing at the Lint step, which meant every step after it
was skipped and had never run on this code: backend pytest, migrations from
scratch, the single-Alembic-head assert, the drift check, and the web build.

Most of this is `ruff --fix` and `eslint` output — import ordering, a
`datetime.UTC` alias, an unused `MATCH_CONTRACT_VERSION` import, a stale
`noqa`. The ones worth a second look:

- The reference-library index warmer swallowed every exception with a bare
  `pass`. An unindexable spectrum really is still a valid entry, but a silent
  `except` also hides real bugs, so it logs now — same idiom `spectra.py`
  already uses for its storage cleanup.
- `SpectrumChart` wrote two refs from the render body. The refs exist to keep
  `zoomRange`/`onZoomChange` out of the draw effect's dependency list, so the
  writes move into an effect declared above the draw and zoom effects —
  effects run in declaration order, so those still read current values, and
  the `useRef` seeds cover the first render.
- `PostGallery` re-sorted its prop into state from an effect. Same sync, done
  during render instead, so there's no cascading second render.

Behaviour is otherwise unchanged: `int(round(x))` drops to `round(x)`,
`markers && markers.length` to `markers?.length`, and `seriesProp!` gives way
to a `seriesProp?.length` check that narrows the type on its own.

+++ backend/app/processing/peaks.py
@@ -16,8 +16,8 @@ row rather than leaving a corpus indexed by two different algorithms.
 """
 from __future__ import annotations
 
-from dataclasses import dataclass, asdict
-from typing import Sequence
+from collections.abc import Sequence
+from dataclasses import asdict, dataclass
 
 import numpy as np
 from scipy.signal import find_peaks, peak_widths
@@ -168,7 +168,7 @@ def detect_peaks(
     # median step — spectra are not guaranteed to be evenly sampled.
     steps = np.diff(x)
     median_step = float(np.median(np.abs(steps))) if steps.size else 0.0
-    distance = max(1, int(round(min_distance_cm1 / median_step))) if median_step > 0 else 1
+    distance = max(1, round(min_distance_cm1 / median_step)) if median_step > 0 else 1
 
     threshold = min_prominence_sigma * sigma
     indices, props = find_peaks(corrected, prominence=threshold, distance=distance)
